fix: allow out-of-order values with a size param on the index entries
The library was designed to have the values and keys in order so that a size
field is not needed on each index entry.

It turns out that it is advantageous to sort the values differently from the
keys in circumstances where loading a chunk of a file would be better optimized
if the adjacent values in the chunk are related to each other.

To allow this, add a size parameter to the index entries.

Signed-off-by: Christian Stewart <christian@aperture.us>